+++ odfdom/src/main/java/org/odftoolkit/odfdom/dom/element/chart/ChartLegendElement.java
@@ -204,7 +204,7 @@ public class ChartLegendElement extends OdfStylableElement {
         (StyleLegendExpansionAspectRatioAttribute)
             getOdfAttribute(OdfDocumentNamespace.STYLE, "legend-expansion-aspect-ratio");
     if (attr != null && !attr.getValue().isEmpty()) {
-      return Double.valueOf(attr.doubleValue());
+      return attr.doubleValue();
     }
     return null;
   }
@@ -221,7 +221,7 @@ public class ChartLegendElement extends OdfStylableElement {
     StyleLegendExpansionAspectRatioAttribute attr =
         new StyleLegendExpansionAspectRatioAttribute((OdfFileDom) this.ownerDocument);
     setOdfAttribute(attr);
-    attr.setDoubleValue(styleLegendExpansionAspectRatioValue.doubleValue());
+    attr.setDoubleValue(styleLegendExpansionAspectRatioValue);
   }
 
   /**
