+++ odfdom/src/main/java/org/odftoolkit/odfdom/dom/element/text/TextAlphabeticalIndexEntryTemplateElement.java
@@ -82,7 +82,7 @@ public class TextAlphabeticalIndexEntryTemplateElement extends OdfStylableElemen
     TextOutlineLevelAttribute attr =
         (TextOutlineLevelAttribute) getOdfAttribute(OdfDocumentNamespace.TEXT, "outline-level");
     if (attr != null && !attr.getValue().isEmpty()) {
-      return Integer.valueOf(attr.intValue());
+      return attr.intValue();
     }
     return null;
   }
@@ -96,7 +96,7 @@ public class TextAlphabeticalIndexEntryTemplateElement extends OdfStylableElemen
   public void setTextOutlineLevelAttribute(Integer textOutlineLevelValue) {
     TextOutlineLevelAttribute attr = new TextOutlineLevelAttribute((OdfFileDom) this.ownerDocument);
     setOdfAttribute(attr);
-    attr.setIntValue(textOutlineLevelValue.intValue());
+    attr.setIntValue(textOutlineLevelValue);
   }
 
   /**
